LayoutTests/tiled-drawing/scrolling/scroll-snap/scroll-snap-async-iframe-expected.txt

Tests that the scroll-snap feature works properly in iframes.

Testing scroll-snap glide for horizontalTarget:
PASS div scrolled to next window.
Testing scroll-snap snap for horizontalTarget:
PASS div honored snap points.
Testing scroll-snap glide for verticalTarget:
PASS div scrolled to next window.
Testing scroll-snap snap for verticalTarget:
PASS div honored snap points.
PASS successfullyParsed is true

TEST COMPLETE
